requirements.txt
# packaging
pip>=9
setuptools>=26

# data science
numpy
pandas
tensorflow-cpu

# utilities
streamlit
opencv-python-headless
patchify
requests
bs4
